clean up json provider implementation
